create point presenter
